requirements.txt
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
PyJWT==1.7.1
pymongo==3.11.3
requests==2.25.1
selenium==3.141.0
soupsieve==2.2
urllib3==1.26.3
Werkzeug==1.0.1
